requirements.txt
numpy
beautifulsoup4
tabula-py
tabulate
requests
